+++ debezium-bom/pom.xml
@@ -29,8 +29,7 @@
         <version.grpc>1.56.1</version.grpc>
 
         <!-- Tracing -->
-        <version.opentelemetry>1.21.0</version.opentelemetry>
-        <version.opentelemetry.alpha>1.22.1-alpha</version.opentelemetry.alpha>
+        <version.opentelemetry.alpha>1.23.0-alpha</version.opentelemetry.alpha>
 
         <!-- HTTP client -->
         <version.okhttp>4.8.1</version.okhttp>
