+++ debezium-interceptor/src/main/java/io/debezium/tracing/DebeziumTracingProducerInterceptor.java
@@ -3,12 +3,15 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.debezium.transforms.tracing;
+package io.debezium.tracing;
 
 import org.apache.kafka.clients.producer.ProducerRecord;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import io.opentelemetry.api.GlobalOpenTelemetry;
 import io.opentelemetry.api.OpenTelemetry;
+import io.opentelemetry.api.internal.ConfigUtil;
 import io.opentelemetry.api.trace.Span;
 import io.opentelemetry.api.trace.SpanKind;
 import io.opentelemetry.api.trace.Tracer;
@@ -16,9 +19,12 @@ import io.opentelemetry.context.Context;
 import io.opentelemetry.context.Scope;
 import io.opentelemetry.context.propagation.TextMapGetter;
 import io.opentelemetry.context.propagation.TextMapPropagator;
-import io.opentelemetry.instrumentation.kafkaclients.TracingProducerInterceptor;
+import io.opentelemetry.instrumentation.kafkaclients.v2_6.TracingProducerInterceptor;
 
 public class DebeziumTracingProducerInterceptor<K, V> extends TracingProducerInterceptor<K, V> {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(DebeziumTracingProducerInterceptor.class);
+    public static final String ARG_OTEL_INSTRUMENTATION_KAFKA_ENABLED = "otel.instrumentation.kafka.enabled";
     private static final OpenTelemetry openTelemetry = GlobalOpenTelemetry.get();
     private static final Tracer tracer = openTelemetry.getTracer(DebeziumTracingProducerInterceptor.class.getName());
     private static final TextMapPropagator TEXT_MAP_PROPAGATOR = openTelemetry.getPropagators().getTextMapPropagator();
@@ -26,8 +32,14 @@ public class DebeziumTracingProducerInterceptor<K, V> extends TracingProducerInt
 
     @Override
     public ProducerRecord<K, V> onSend(ProducerRecord<K, V> producerRecord) {
-        Context parentContext = TEXT_MAP_PROPAGATOR.extract(Context.current(), producerRecord, GETTER);
+        if (isInstrumentationKafkaEnabled()) {
+            LOGGER.warn(
+                    "To enable end-to-end traceability with Debezium you need to disable automatic Kafka instrumentation. To disable, run your JVM with -D{}=false\"",
+                    ARG_OTEL_INSTRUMENTATION_KAFKA_ENABLED);
+            return producerRecord;
+        }
 
+        Context parentContext = TEXT_MAP_PROPAGATOR.extract(Context.current(), producerRecord, GETTER);
         Span interceptorSpan = tracer.spanBuilder("onSend")
                 .setSpanKind(SpanKind.INTERNAL)
                 .setParent(parentContext)
@@ -39,6 +51,9 @@ public class DebeziumTracingProducerInterceptor<K, V> extends TracingProducerInt
         finally {
             interceptorSpan.end();
         }
+    }
 
+    private static boolean isInstrumentationKafkaEnabled() {
+        return Boolean.parseBoolean(ConfigUtil.getString(ARG_OTEL_INSTRUMENTATION_KAFKA_ENABLED, "true"));
     }
 }
