+++ pom.xml
@@ -176,6 +176,8 @@
         <module>debezium-connect-rest-extension</module>
         <module>debezium-schema-generator</module>
         <module>debezium-storage</module>
+        <module>debezium-interceptor</module>
+
     </modules>
 
     <distributionManagement>
